bmad_system/bmad_backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-Migrate==4.0.5
pyyaml==6.0.2
python-docx==1.2.0
PyMuPDF==1.26.3
pandas==2.3.1
openpyxl==3.1.5
python-pptx==1.0.2
pillow==11.3.0
pytesseract==0.3.13
google-generativeai==0.8.5
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.1

